fix cueq tests
